refactor: change of project files structure, move functions to separate files as utility functions
